requirements.txt
tensorflow
keras
numpy
opencv-python
Pillow
